+++ src/lib_dal_node/daemon.ml
@@ -280,7 +280,7 @@ let update_and_register_profiles ctxt =
   let*! () = Node_context.set_profile_ctxt ctxt profile_ctxt in
   return_unit
 
-let run ?(disable_shard_validation = false) ~ignore_pkhs ~data_dir
+let run ?(disable_shard_validation = false) ~ignore_pkhs ~data_dir ~config_file
     ~configuration_override () =
   let open Lwt_result_syntax in
   let*! () =
@@ -309,14 +309,14 @@ let run ?(disable_shard_validation = false) ~ignore_pkhs ~data_dir
           ignore_l1_config_peers;
           _;
         } as config) =
-    let*! result = Configuration_file.load ~data_dir in
+    let*! result = Configuration_file.load ~config_file in
     match result with
     | Ok configuration -> return (configuration_override configuration)
     | Error _ ->
-        let*! () = Event.emit_data_dir_not_found ~path:data_dir in
+        let*! () = Event.emit_config_file_not_found ~path:config_file in
         (* Store the default configuration if no configuration were found. *)
         let configuration = configuration_override Configuration_file.default in
-        let* () = Configuration_file.save configuration in
+        let* () = Configuration_file.save ~config_file configuration in
         return configuration
   in
   let*! () = Event.emit_configuration_loaded () in